feat: Add settings for display name and share template
Introduces new settings for `myDisplayName` and `shareTemplate`. The `myDisplayName` allows users to set a preferred name for themselves within the app, and `shareTemplate` enables customization of the message sent to participants regarding their outstanding bills.

Also, the `CreateBill` and `BillDetails` components are updated to accept and utilize these new settings. `BillDetails` now takes all `bills` as a prop for improved context when generating share messages.